file-parser-demo/ds-plugin-interf/src/main/java/com/hillstone/dsgp/asset/datasource/plugin/interf/common/util/ObjectUtils.java
package com.hillstone.dsgp.asset.datasource.plugin.interf.common.util;

import java.lang.reflect.Array;
import java.util.Collection;
import java.util.Map;
import java.util.Optional;

/**
 * ObjectUtils
 *
 * @author wangjian2
 * @date 2024/4/16 13:40
 */
public class ObjectUtils {

    /**
     * 对象判空
     *
     * @see org.springframework.util.ObjectUtils#isEmpty(Object)
     * @param obj
     * @return
     */
    public static boolean isEmpty(Object obj) {
        if (obj == null) {
            return true;
        }

        if (obj instanceof Optional) {
            return !((Optional<?>) obj).isPresent();
        }
        if (obj instanceof CharSequence) {
            return ((CharSequence) obj).length() == 0;
        }
        if (obj.getClass().isArray()) {
            return Array.getLength(obj) == 0;
        }
        if (obj instanceof Collection) {
            return ((Collection<?>) obj).isEmpty();
        }
        if (obj instanceof Map) {
            return ((Map<?, ?>) obj).isEmpty();
        }

        // else
        return false;
    }
}
